Add schema generation for CRD
The main difficulty is that kubernetes doesn't support $ref keys in the
openapi spec for CRDs, so they need to be completely flattened into a
single document.

Here, we do it by applying controller-gen to a Go type definition. I'm
not sure this is ideal, since it requires an entire Go toolchain, and
doesn't inject useful things like patch merge keys, but it works well
enough to get started.