Handle terminating pods correctly
This required a bit of a rework of the CRD's data structures, since we
need to patch CRD's status from both the pod termination handler and the
CRD's update handler.

I think the way I've implemented this is well thought-through in terms
of data races. It at least seems to pass the smoke-tests.