Track not-yet-created pods
It turns out there was a race condition in the pod handler - the main
cluster handler would trigger itself by updating its own status. The pod
index was not always updated between the two consecutive runs of this
handler.

This adds a set that tracks the created pods that haven't yet had their
creation event handled by the cluster, and if there are any, simply
defers the main cluster handler for a short delay.