Make scale-down test less flaky
The scale down test was doing a lot of work in its test function (wait
for workers to be ready, scale cluster, wait for worker to be deleted).
This was making it a bit flaky.

Increasing the timeout helped, but moving the "wait for worker to be
ready" to a setup phase is cleaner.