Initial operator release (Go version) (#2)
* Operator skeleton

This is a no-op operator, but it includes code generation for the CRD
type, and a skeleton for test setup.

Based mostly on github.com/kubernetes/sample-operator.

* Sketch out listeners

This finishes the input phase - we listen for changes to Dask clusters
and created resources (pods & deployments).

* Add scheduler templating logic & tests

This required a bit of a refactor of the controller, to expose the
informers so that fake values can be injected during tests.

* Create worker pods

* Additional tests

The label selector for pods was incorrect, fixed now.

* Handle retiring workers

For now, this requests that Dask shuts down workers, but does not
enforce this.

* Add e2e test

The E2E test suite requires an existing kind cluster. I wasn't able to
get the automatic setup of a kind cluster working, especially the image
upload.

* Separate template functions

* Test & fix worker template generation

Turns out that go loops loop over values, not references. Use the index
to return a reference.

* E2E test scheduler pod creation

* Add CI

* Test that pods aren't retired twice

* Track worker retire time

We don't do anything with this yet, but we will want to force-kill
workers when they have been told to retire but haven't exited after some
time.

* Reduce go-install commands in Dockerfile

* Refactor test fixtures

* Namespaced operation support

This is useful for testing, because the roles can be cleaned up
automatically.

* Test & fix scheduler labels

* Check workers are ready

Not just running. This requires injecting the port for health-checks,
which must be done via the CLI. Maybe we should inject these args
unconditionally, but what if pods are deployed with wrappers?

* Add subresource annotations

And make sure that the status/spec fields amtch up

* Fix service informer syncing

* Test and fix pod retiring

* Support explicit kubeconfig for debugging

* Fix status updates

The different applyconfiguration patches need to have different field
managers.

* Fix pod updates

Due to an inverted condition, we were only processing new, unowned
resources, so we missed pod deletions.

* Clean up completed pods

* Test that workers are replaced

* Add .dockerignore

* Test cluster state instead of actions

This is much easier to keep consistent

* Fix fixture labels

* Fix integration test kind cluster

* Fix scale down behaviour

* Expose prometheus metrics

* Add file-level docstrings

* Add cluster deletion handler

This is a no-op, deleted clusters don't trigger any controller logic, as
we rely on kubernetes GC to clean up their components. But add the
handler anyway, to avoid confusion.

* Make scale-down test less flaky

The scale down test was doing a lot of work in its test function (wait
for workers to be ready, scale cluster, wait for worker to be deleted).
This was making it a bit flaky.

Increasing the timeout helped, but moving the "wait for worker to be
ready" to a setup phase is cleaner.